rand

+++ sim.h
@@ -23,8 +23,8 @@ class Sim{
     void print_users();
     void print_Uservector(std::vector<User> printing);
     void print_adjMatrix();
-    int averagePing();
-    int calculate_ping(std::string input);
+    double averagePing();
+    double calculate_ping(std::string input);
 
   private:
     Server server;
